# server/docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.AIAISummaryRequest:
    properties:
      content:
        description: 文章内容
        type: string
    required:
    - content
    type: object
  dto.AIAISummaryResponse:
    properties:
      summary:
        description: 生成的AI摘要
        type: string
    type: object
  dto.AISummaryRequest:
    properties:
      content:
        description: 文章内容
        type: string
    required:
    - content
    type: object
  dto.AISummaryResponse:
    properties:
      summary:
        description: 生成的摘要
        type: string
    type: object
  dto.AITestRequest:
    properties:
      api_key:
        type: string
      base_url:
        type: string
      model:
        type: string
    required:
    - api_key
    - base_url
    - model
    type: object
  dto.AITitleRequest:
    properties:
      content:
        description: 文章内容
        type: string
    required:
    - content
    type: object
  dto.AITitleResponse:
    properties:
      title:
        description: 生成的标题
        type: string
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
info:
  contact:
    email: talen2004@163.com
    name: M.Talen
  description: 一个基于 Go 语言的现代化博客后端服务
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Flec-Server
  version: v1
paths:
  /admin/ai/ai-summary:
    post:
      consumes:
      - application/json
      description: 基于文章内容和提示词生成AI摘要
      parameters:
      - description: 文章内容
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AIAISummaryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.AIAISummaryResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 生成AI摘要
      tags:
      - AI功能
  /admin/ai/summary:
    post:
      consumes:
      - application/json
      description: 基于文章内容和提示词生成文章摘要
      parameters:
      - description: 文章内容
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AISummaryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.AISummaryResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 生成文章摘要
      tags:
      - AI功能
  /admin/ai/test:
    post:
      consumes:
      - application/json
      description: 使用提供的配置测试AI服务连通性
      parameters:
      - description: AI配置
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AITestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 测试AI配置
      tags:
      - AI功能
  /admin/ai/title:
    post:
      consumes:
      - application/json
      description: 基于文章内容和提示词生成标题
      parameters:
      - description: 文章内容
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AITitleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.AITitleResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 生成标题
      tags:
      - AI功能
...
securityDefinitions:
  BearerAuth:
    description: 在请求头中添加 Bearer Token，格式：Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"